removed some compiler warnings
git-svn-id: https://subversion.assembla.com/svn/zappralot/trunk/libfortres@762 c37337d0-301c-4318-a67e-7f75229f9ee7

+++ fort_dirent.cpp
@@ -188,7 +188,7 @@ make_realpath( const char *filePath, char *buff, size_t len )
   size_t tgtLen = 0;
 
 #if defined _MSC_VER
-  tgtLen = GetFullPathName( filePath, len, buff, NULL );
+  tgtLen = GetFullPathName( filePath, static_cast<DWORD>(len), buff, NULL );
 #else
   char *ptr;
